+++ scm-ui/ui-components/src/forms/TriStateCheckbox.tsx
@@ -21,31 +21,52 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-import React from "react";
-import classNames from "classnames";
+import React, { FC } from "react";
+import Icon from "../Icon";
 
 type Props = {
-  title?: string;
-  iconStyle: string;
-  name: string;
-  color: string;
-  className?: string;
-  onClick?: () => void;
+  onChange?: (value: boolean) => void;
+  checked: boolean;
+  indeterminate?: boolean;
+  disabled?: boolean;
+  label?: string;
 };
 
-export default class Icon extends React.Component<Props> {
-  static defaultProps = {
-    iconStyle: "fas",
-    color: "grey-light"
+const TriStateCheckbox: FC<Props> = ({ onChange, checked, indeterminate, disabled, label }) => {
+  const onCheckboxChange = () => {
+    if (onChange) {
+      console.log("clickediclick");
+      onChange(!checked);
+    }
   };
 
-  render() {
-    const { title, iconStyle, name, color, className, onClick } = this.props;
-    if (title) {
-      return (
-        <i onClick={onClick} title={title} className={classNames(iconStyle, "fa-fw", "fa-" + name, `has-text-${color}`, className)} />
-      );
-    }
-    return <i onClick={onClick} className={classNames(iconStyle, "fa-" + name, `has-text-${color}`, className)} />;
+  let icon;
+  if (indeterminate) {
+    icon = "minus-square";
+  } else if (checked) {
+    icon = "check-square";
+  } else {
+    icon = "square";
   }
-}
+
+  let className;
+  if (!checked || indeterminate) {
+    className = "far";
+  } else {
+    className = "fa";
+  }
+
+  let color;
+  if (disabled) {
+    color = "grey-light";
+  } else if (checked || indeterminate) {
+    color = "blue";
+  } else {
+    color = "black";
+  }
+
+  return <><Icon iconStyle={"is-outlined"} name={icon} className={className} color={color} />{" "}
+  {label}</>;
+};
+
+export default TriStateCheckbox;